Implement the `bytes` command
The `bytes` command on HP calculators returns the size of the object
in the first level of the stack, and the CRC32 (hash) of the object
in the second level fo the stack.

The implememntation here returns the _complete_ representation of the
object in level 2. Also, the size is always an integral number of
bytes (the Saturn-based calculators were 4-bit, so byte sizes could be
half integer)

The lowest part of the number matches the result of the `type` command.

Fixes: #284

+++ src/settings.h
@@ -175,8 +175,8 @@ public:
     uint16_t standard_exp;      // Maximum exponent before switching to sci
     angles   angle_mode;        // Angle mode ( degrees, radians or grads)
     uint8_t  base;              // The default base for #numbers
-    uint16_t wordsize;          // Wordsize for binary numbers (in bits)
-    uint16_t maxbignum;         // Maximum size for a bignum (in bits)
+    size_t   wordsize;          // Wordsize for binary numbers (in bits)
+    size_t   maxbignum;         // Maximum size for a bignum (in bits)
     uint16_t maxrewrites;       // Maximum number of rewrites
     uint16_t fraciter;          // Number of iterations for ->Q
     uint16_t fracprec;          // Number of digits for ->Q
